backend/requirements.txt
Django==5.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
Pillow==10.1.0